+++ vojuck.victoria/P4/findMaxSum.hpp
@@ -1,0 +1,6 @@
+#ifndef FINDMAXSUM_HPP
+#define FINDMAXSYM_HPP
+
+int findMaxSum(int* array, int rows, int cols);
+
+#endif
