+++ vojuck.victoria/P4/inputMatrix.cpp
@@ -1,0 +1,16 @@
+#include "inputArray.hpp"
+#include <iostream>
+#include <algorithm>
+
+size_t inputArray(std::istream & in, int * array, size_t size, size_t toread)
+{
+  for (size_t i = 0; i < std::min(size, toread); ++i)
+  {
+    in >> array[i];
+    if (!in)
+    {
+      return i;
+    }
+  }
+  return std::min(size, toread);
+}
