+++ vojuck.victoria/P4/inputMatrix.hpp
@@ -1,0 +1,6 @@
+#ifndef INPUTMATRIX_HPP
+#define INPUTMATRIX_HPP
+
+size_t inputArray(std::istream & in, int * array, size_t size, size_t toread);
+
+#endif
