+++ vojuck.victoria/P4/isUpperTriangular.hpp
@@ -1,0 +1,6 @@
+#ifndef ISUPPERTRIANGULAR_HPP
+#define ISUPPERTRIANGULAR_HPP
+
+bool isUpperTriangular(int* matrix, int rows, int cols);
+
+#endif
